Move the date helpers out of CleanFaresCommand
startOfYear and getFirstDateAfter were exported from the command module
only so the tests could reach them - nothing outside CleanFaresCommand
imported either one. They are general PlainDate operations, so they move
to PlainDate.ts alongside dayOfWeek and the comparison helpers, and the
command imports them like any other caller.

The tests move with them into PlainDate.spec.ts, which leaves
test/cli/CleanFaresCommand.spec.ts empty, so it is removed.
CleanFaresCommand now exports only the command class.
